# openapi.yaml
openapi: 3.0.1
info:
  description: "HalykAcademy final project Bookstore API definition"
  title: BookStore API definiton
  version: v1
servers:
  - url: http://localhost:8081/api
    description: Generated server url
paths:
  /users:
    get:
      tags:
        - users
      summary: "Gets list of all users"
      operationId: showAllUsers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
    put:
      tags:
        - users
      summary: "Updates a user"
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPutDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserPutDto'
    post:
      tags:
        - users
      summary: "Adds a new user"
      operationId: addNewUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPostDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserPostDto'
...
  /users/{id}:
    get:
      tags:
        - users
      summary: "Gets a user by ID"
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
    delete:
      tags:
        - users
      summary: "Deletes a user by ID"
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
components:
  schemas:
    UserPostDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        login:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
        blocked:
          type: boolean
...
    UserPutDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        login:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
        blocked:
          type: boolean
...
    OrderSlimDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        login:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN
        orders:
          type: array
          items:
            $ref: '#/components/schemas/OrderSlimDto'
        blocked:
          type: boolean
...